Run email sending in child process
