+++ next-email-app/package.json
@@ -27,6 +27,7 @@
     "files": [
       "index.js",
       "preload.js",
+      "sendEmail.js",
       "out/**/*"
     ]
   },
